Move conformance.go from blob package to driver/drivertest package.
